rc: set time on boot
We need the time to be roughly correct when we call wpa_supplicant.
Since we don't have an RTC, we save the time in "nvram" in a "date"
variable.  The "date" variable is saved in seconds from epoch so
it can easily be passed to stime().

if the nvram "date" variable isn't set we expect nvram_get_int to
return 0 and thus we'll skip calling stime().

We could do this in /jffs/scripts/init-start, but this is slightly
more "permant".

The init-start script would look something like:

	# Check if we have nvram 'date' set, if so convert to MMDDhhmmYYYY
	# format from seconds from epoch that date command expects.
	nvdate=$(/bin/nvram get date)
	if [ $nvdate ]; then
	   date `date --date=@${nvdate} +%m%d%H%M%Y`
	fi

	# Install cronjob to set "nvram" date once a hour
	/usr/sbin/cru a nvramdate "1 * * * * /bin/nvram set date=\`/bin/date +%s\` && /bin/nvram commit"

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>