ddns: implement support for DYNV6.com
For dynv6.com to work correctly for setting DNS A record we need to
utilize the DynDNS API that is supported.  inadyn can handle this
via the 'custom' config (instead of dynv6 provider config).  The
config for inadyn looks like:

custom dynv6 {
	ddns-server = dynv6.com
	ddns-path = "/nic/update?hostname=%h&myip=%i"
	...
}

Update rc/services.c to handle this custom config, which is similar
to what is needed for WWW.NAMECHEAP.COM.

Signed-off-by: Kumar Gala <galak@kernel.org>